+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - change - do not override existing bit.json on bit init
 - backward compatibility for components with environments with latest version
 - fix unhandled rejection on bit test
+- show dependent component id when trying to install missing environment
 - change no-launch-browser to suppress-browser-launch in bit login flag
 
 ## [13.0.0-dev.11] - 2018-06-13